+++ src/game.js
@@ -1,1 +1,0 @@
-window.onload = () => console.log("hello world2356789")